FileType.jl

Documentation for FileTypes.jl

Documentation

FileTypes.Is — Method

API to check if it is of given type or not.
Args:
    FileType.FileTypes (object)  
    filename (String) i.e filepath 
Returns:
    true if the filetype matches with the given object else flase

FileTypes.is_extension_supported — Method

Checks if the given extension string is
one of the supported by the library.
Args:
    ext (String): file extension string. E.g: jpg, png, mp4, mp3
Returns:
    true if the file extension is supported.
    Otherwise false.

FileTypes.matcher — Method

API to return FileType.Type object which has files as extension of the file and MIME of the file
Args:
    filename (String) i.e filepath 
Returns:
    FileType.Type object